Add request cancellation, add infinite decimals

+++ src/FrontEASE.DataContracts/Models/Core/Tasks/Data/Configs/Modules/Values/TaskModuleParameterValueCoreDto.cs
@@ -6,7 +6,7 @@ namespace FrontEASE.DataContracts.Models.Core.Tasks.Data.Configs.Modules.Values
     {
         #region Data
 
-        public float? FloatValue { get; set; }
+        public double? FloatValue { get; set; }
         public int? IntValue { get; set; }
         public string? StringValue { get; set; }
         public bool? BoolValue { get; set; }
